Temporarily override the get method for the PyG datasets
Temporarily override the get method for the datasets provided by PyTorch-Geometric to make the AutoGL compatible with PyTorch-Geometric>=1.6.2.
For version 1.6.2 or higher, the PyG InMemoryDataset has an additional internal attribute __data_list__ by default. If an in-memory dataset has the attribute and it is not None, then the item in the given specific index will be directly returned without accessing the data property.
Besides, the PyG matainer rusty1s also mentioned that "accessing dataset.data is not recommended" in https://github.com/rusty1s/pytorch_geometric/issues/1984.
In other words, mutating the data property of InMemoryDataset is not recommended by the PyTorch-Geometric. However, the current version of AutoGL does modify the data property of dataset.
Thus the get method of the PyG provided datasets is overrided to delete the __data_list__ attribute in advance before calling the virtual get method, aiming to temporarily fix the compatibility issue.

In addition, the ModelNet10 and ModelNet40 datasets are significantly modified.