Attempt to design an experimental abstract TargetDependantSampler
Attempt to design an experimental abstract TargetDependantSampler interface for Node-wise Sampling and Layer-wise Sampling.
Refactor the Neighbor Sampler and Layer Dependant Importance Sampling (LADIES) to be the implementations of the TargetDependantSampler interface, meanwhile fix a sampling issue, i.e. Samplers and the corresponding Trainers only sample edge_index without sampling the involved sub-graph and remapping the node indexes for the sampled sub-graph previously. Now Both the Neibor sampler and the LADIES Sampler will sample a sub-graph composed of all the nodes involved in the sampled edges for all layers, according to the sampled set of edge_index.
TODO: The LADIES Sampler MUST be fixed recently, Zixin plans to do it.
TODO: The unified abstraction of autogl.data.Data and autogl.data.Dataset is urgently necessary for the long-term development of the AutoGL library. Schedule to do the abstraction for next upcoming major version.